+++ tests/conftest.py
@@ -9,7 +9,9 @@ Utility functions that are used to configure Py.Test context.
 
 """
 
+import abc
 import os
+
 import pytest
 
 
@@ -61,3 +63,31 @@ def state(request):
         setup = lambda: State(),
         scope = "session"
     )
+
+
+def _add_metaclass(metaclass):
+    """Class decorator for creating a class with a metaclass."""
+    def wrapper(cls):
+        orig_vars = cls.__dict__.copy()
+        slots = orig_vars.get('__slots__')
+        if slots is not None:
+            if isinstance(slots, str):
+                slots = [slots]
+            for slots_var in slots:
+                orig_vars.pop(slots_var)
+        orig_vars.pop('__dict__', None)
+        orig_vars.pop('__weakref__', None)
+
+        return metaclass(cls.__name__, cls.__bases__, orig_vars)
+
+    return wrapper
+
+
+@_add_metaclass(abc.ABCMeta)
+class MetaClassParentType(object):
+    """The parent meta class object."""
+    pass
+
+class MetaClassType(MetaClassParentType):
+    """The child meta class type."""
+    pass